feat: :sparkles: Basic Setup and Login flow
Implementing a basic setup of backend (java / spring boot) and a
frontend usint typescript and react. The skeleton is pretty rough at the
moment, but features basic authentication flows using HTTP Only Cookies
and frontend - backend communication.

Focus on this basic setup is, to get the authentication, framework and
libary setup done, to down the line speed up development.